+++ decidim-admin/app/controllers/decidim/admin/newsletters_controller.rb
@@ -7,9 +7,10 @@ module Decidim
       include Decidim::NewslettersHelper
       include Decidim::Admin::NewslettersHelper
       include Paginable
+      helper_method :newsletter
 
       def index
-        enforce_permission_to :read, :newsletter
+        enforce_permission_to :index, :newsletter
         @newsletters = collection.order(Newsletter.arel_table[:created_at].desc)
         @newsletters = paginate(@newsletters)
       end
@@ -20,16 +21,14 @@ module Decidim
       end
 
       def show
-        @newsletter = collection.find(params[:id])
-        @email = NewsletterMailer.newsletter(current_user, @newsletter)
-        enforce_permission_to :read, :newsletter, newsletter: @newsletter
+        enforce_permission_to :read, :newsletter, newsletter: newsletter
+        @email = NewsletterMailer.newsletter(current_user, newsletter)
       end
 
       def preview
-        @newsletter = collection.find(params[:id])
-        enforce_permission_to :read, :newsletter, newsletter: @newsletter
+        enforce_permission_to :read, :newsletter, newsletter: newsletter
 
-        email = NewsletterMailer.newsletter(current_user, @newsletter)
+        email = NewsletterMailer.newsletter(current_user, newsletter)
         Premailer::Rails::Hook.perform(email)
         render html: email.html_part.body.decoded.html_safe
       end
@@ -53,17 +52,15 @@ module Decidim
       end
 
       def edit
-        @newsletter = collection.find(params[:id])
-        enforce_permission_to :update, :newsletter, newsletter: @newsletter
-        @form = form(NewsletterForm).from_model(@newsletter)
+        enforce_permission_to :update, :newsletter, newsletter: newsletter
+        @form = form(NewsletterForm).from_model(newsletter)
       end
 
       def update
-        @newsletter = collection.find(params[:id])
-        enforce_permission_to :update, :newsletter, newsletter: @newsletter
+        enforce_permission_to :update, :newsletter, newsletter: newsletter
         @form = form(NewsletterForm).from_params(params)
 
-        UpdateNewsletter.call(@newsletter, @form, current_user) do
+        UpdateNewsletter.call(newsletter, @form, current_user) do
           on(:ok) do |newsletter|
             flash[:notice] = I18n.t("newsletters.update.success", scope: "decidim.admin")
             redirect_to action: :show, id: newsletter.id
@@ -78,10 +75,9 @@ module Decidim
       end
 
       def destroy
-        @newsletter = collection.find(params[:id])
-        enforce_permission_to :destroy, :newsletter, newsletter: @newsletter
+        enforce_permission_to :destroy, :newsletter, newsletter: newsletter
 
-        DestroyNewsletter.call(@newsletter, current_user) do
+        DestroyNewsletter.call(newsletter, current_user) do
           on(:already_sent) do
             flash.now[:error] = I18n.t("newsletters.destroy.error_already_sent", scope: "decidim.admin")
             redirect_to :back
@@ -95,17 +91,15 @@ module Decidim
       end
 
       def select_recipients_to_deliver
-        @newsletter = collection.find(params[:id])
-        enforce_permission_to :update, :newsletter, newsletter: @newsletter
-        @form = form(SelectiveNewsletterForm).from_model(@newsletter)
+        enforce_permission_to :update, :newsletter, newsletter: newsletter
+        @form = form(SelectiveNewsletterForm).from_model(newsletter)
       end
 
       def deliver
-        @newsletter = collection.find(params[:id])
-        enforce_permission_to :update, :newsletter, newsletter: @newsletter
+        enforce_permission_to :update, :newsletter, newsletter: newsletter
         @form = form(SelectiveNewsletterForm).from_params(params)
 
-        DeliverNewsletter.call(@newsletter, @form, current_user) do
+        DeliverNewsletter.call(newsletter, @form, current_user) do
           on(:ok) do
             flash[:notice] = I18n.t("newsletters.deliver.success", scope: "decidim.admin")
             redirect_to action: :index
@@ -126,7 +120,11 @@ module Decidim
       private
 
       def collection
-        Newsletter.where(organization: current_organization)
+        @collection ||= Newsletter.where(organization: current_organization)
+      end
+
+      def newsletter
+        @newsletter ||= collection.find_by(id: params[:id])
       end
     end
   end